requirements.txt
markdown == 3.2.1
django_taggit == 1.2.0
psycopg2-binary
